+++ lib/moqx/protocol/resolver.ex
@@ -10,8 +10,6 @@ defmodule MOQX.Protocol.Resolver do
   alias MOQX.Protocol
 
   @built_ins %{
-    cloudflare_draft_14: MOQX.Protocol.CloudflareDraft14,
-    draft_16: MOQX.Protocol.Draft16,
     draft_18: MOQX.Protocol.Draft18,
     moq_lite_05: MOQX.Protocol.MOQLite05
   }